# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/user/balance": {
            "get": {
                "description": "用户获取积分余额",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户获取积分余额",
                "responses": {
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMsgBody"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "用户通过邮箱、密码和 Google 验证码进行登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功，返回用户信息及 token",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRsp"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMsgBody"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "description": "用户登出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户登出",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRsp"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMsgBody"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功，返回用户信息及 token",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRsp"
                        }
                    },
                    "400": {
                        "description": "登录失败",
                        "schema": {
                            "$ref": "#/definitions/rsp.ErrorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ErrorMsgBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.LoginReq": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 6
                }
            }
        },
        "dto.LoginRsp": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "perm": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
        "dto.RegisterReq": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 32
                },
                "invite_code": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 6
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "invite_code": {
                    "type": "string"
                },
                "invite_points": {
                    "type": "number"
                },
                "invited_by": {
                    "type": "string"
                },
                "key": {
                    "description": "todo key 不能返回给用户",
                    "type": "string"
                },
                "last_balance": {
                    "type": "number"
                },
                "last_check_at": {
                    "type": "string"
                },
                "last_login_at": {
                    "type": "string"
                },
                "last_login_ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "recharge_points": {
                    "type": "number"
                },
                "role": {
                    "$ref": "#/definitions/role.RoleType"
                },
                "secret": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/status.UserStatus"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "role.RoleType": {
            "type": "string",
            "enum": [
                "admin",
                "user"
            ],
            "x-enum-varnames": [
                "Admin",
                "User"
            ]
        },
        "rsp.ErrorBody": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "status.UserStatus": {
            "type": "string",
            "enum": [
                "1",
                "0",
                "-1"
            ],
            "x-enum-varnames": [
                "UserOn",
                "UserOff",
                "UserNone"
            ]
        }
    }
}